Finished the Scene
Created enemies and ways to move scene using a Teleporter Script.

Created an object key to become one of the ways to open a door. Create an inventory UIPrefab.

Create a teleport to move from position A to destination B. Also, teleport to move to another scene (explained below)

Changed the first scene from the game, to the scene I made. From the scene I made, it will connect to the pre-built scenes.